backend_llm/requirements.txt
google-generativeai
load_dotenv
flask
llama-index
flask-cors
